Remove ItemSize from ViewModel and inject it into mapper
