+++ CodingSession/SceneDelegate.swift
@@ -17,8 +17,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let window = UIWindow(windowScene: windowScene)
         
         let vm = GalleryViewModel(videoProvider: VideoProvider(),
-                                  cellWidth: UIConstants.itemSize.width,
-                                  mapper: AssetMapper(manager: .default()).map(asset:imageSize:))
+                                  mapper: AssetMapper(manager: .default(), imageSize: UIConstants.itemSize.width).map(asset:))
         
         window.rootViewController = ViewController(viewModel: vm)
         window.makeKeyAndVisible()
